refactor: 안쓰는 메소드 정리
